+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/black-forest-labs
 
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [db913bd]
+  - @ai-sdk/provider@3.0.0-beta.17
+  - @ai-sdk/provider-utils@4.0.0-beta.34
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- 3922a5f: feat(provider/black-forest-labs): make polling timeout configurable
+
 ## 1.0.0-beta.0
 
 ### Major Changes