+++ packages/perplexity/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/perplexity
 
+## 3.0.0-beta.36
+
+### Patch Changes
+
+- Updated dependencies [db913bd]
+  - @ai-sdk/provider@3.0.0-beta.17
+  - @ai-sdk/provider-utils@4.0.0-beta.34
+
+## 3.0.0-beta.35
+
+### Patch Changes
+
+- e623580: feat (provider/perplexity): add reasoning tokens
+
 ## 3.0.0-beta.34
 
 ### Patch Changes